+++ changelog.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- 2023-08-03
+  - Adding highlighted options connectors (on mouse over).
 - 2023-08-02
   - Fixing drag amount when moving boxes.
   - Drawing connections more beautiful with cubic Bézier curves.